load_balancer/requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
sqlalchemy==2.0.21
pathlib
pyodbc==4.0.39
prometheus_flask_exporter~=0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-crontab==3.0.0
docker==6.1.3
